+++ example/storybook/src/components/Disclosure/Actionsheet/ActionsheetVirtualizedList.tsx
@@ -76,6 +76,7 @@ const ActionsheetWithVirtualizedList = ({
           <ActionsheetDragIndicator />
         </ActionsheetDragIndicatorWrapper>
         <ActionsheetVirtualizedList
+          h="$56"
           data={data}
           initialNumToRender={5}
           renderItem={({ item }: any) => <Item title={item.title} />}