+++ example/storybook/src/getting-started/Installation/Stackoverflow.tsx
@@ -5,8 +5,8 @@ function Stackoverflow() {
   const { colorMode } = useContext(LayoutContext);
   return colorMode === 'light' ? (
     <svg
-      width="21"
-      height="24"
+      width={21}
+      height={24}
       viewBox="0 0 21 24"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
@@ -18,8 +18,8 @@ function Stackoverflow() {
     </svg>
   ) : (
     <svg
-      width="21"
-      height="24"
+      width={21}
+      height={24}
       viewBox="0 0 21 24"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"