+++ example/storybook/src/react-native-components/ScrollView/ScrollView.tsx
@@ -1,1 +1,84 @@
-export { ScrollView } from '@gluestack-ui/themed';
+import {
+  ScrollView,
+  Center,
+  Heading,
+  VStack,
+  config,
+  Text,
+} from '@gluestack-ui/themed';
+import React from 'react';
+
+export default function ScrollViewStory() {
+  return (
+    <Center>
+      <ScrollView h="$80" w="$80">
+        <Center mt="$3" mb="$4">
+          <Heading fontSize="xl">Primary</Heading>
+        </Center>
+        <VStack flex={1}>
+          {Object.keys(config.theme.tokens.colors).map((key) => {
+            if (
+              key.includes('primary') &&
+              !(
+                key.includes('600') ||
+                key.includes('700') ||
+                key.includes('800') ||
+                key.includes('900') ||
+                key.includes('950')
+              )
+            )
+              return (
+                <Center py="$4" bg={config.theme.tokens.colors[key]}>
+                  <Text>{key}</Text>
+                </Center>
+              );
+          })}
+        </VStack>
+        <Center mt="$10" mb="$4">
+          <Heading fontSize="$xl">Yellow</Heading>
+        </Center>
+        <VStack flex="1">
+          {Object.keys(config.theme.tokens.colors).map((key) => {
+            if (
+              key.includes('yellow') &&
+              !(
+                key.includes('600') ||
+                key.includes('700') ||
+                key.includes('800') ||
+                key.includes('900') ||
+                key.includes('950')
+              )
+            )
+              return (
+                <Center py="$4" bg={config.theme.tokens.colors[key]}>
+                  <Text>{key}</Text>
+                </Center>
+              );
+          })}
+        </VStack>
+        <Center mt="$10" mb="$4">
+          <Heading fontSize="$xl">Violet</Heading>
+        </Center>
+        <VStack flex="1">
+          {Object.keys(config.theme.tokens.colors).map((key) => {
+            if (
+              key.includes('violet') &&
+              !(
+                key.includes('600') ||
+                key.includes('700') ||
+                key.includes('800') ||
+                key.includes('900') ||
+                key.includes('950')
+              )
+            )
+              return (
+                <Center py="$4" bg={config.theme.tokens.colors[key]}>
+                  <Text>{key}</Text>
+                </Center>
+              );
+          })}
+        </VStack>
+      </ScrollView>
+    </Center>
+  );
+}