+++ packages/themed/src/components/Alert/styled-components/Icon.tsx
@@ -1,4 +1,4 @@
-import { styled } from '../../styled';
+import { styled } from '@gluestack-style/react';
 import { StyledIcon } from '../../Icons/styled-components';
 
 export default styled(