+++ packages/themed/src/components/Image/styled-components/Root.tsx
@@ -1,11 +1,10 @@
 import { Image } from 'react-native';
-import { styled } from '../../styled';
+import { styled } from '@gluestack-style/react';
 
 export default styled(
   Image,
   {
     maxWidth: '$full',
-
     variants: {
       size: {
         '2xs': {