+++ packages/themed/src/components/Select/styled-components-actionsheet/Content.tsx
@@ -1,5 +1,6 @@
 import { Motion } from '@legendapp/motion';
-import { styled } from '../../styled';
+import { styled } from '@gluestack-style/react';
+import { AnimationResolver } from '@gluestack-style/animation-plugin';
 
 export default styled(
   Motion.View,
@@ -7,18 +8,19 @@ export default styled(
     alignItems: 'center',
     borderTopLeftRadius: '$3xl',
     borderTopRightRadius: '$3xl',
-    maxHeight: '70%',
     p: '$2',
     bg: '$backgroundLight0',
     _sectionHeaderBackground: {
       bg: '$backgroundLight0',
     },
+    // @ts-ignore
     _dark: {
       bg: '$backgroundDark900',
       _sectionHeaderBackground: {
         bg: '$backgroundDark900',
       },
     },
+    // @ts-ignore
     _web: {
       userSelect: 'none',
     },
@@ -29,5 +31,8 @@ export default styled(
   {
     componentName: 'SelectActionsheetContent',
     descendantStyle: ['_sectionHeaderBackground'],
-  } as const
+  } as const,
+  {
+    plugins: [new AnimationResolver({})],
+  }
 );